linux support progress

+++ kernels/makeSphere.cl
@@ -5,8 +5,8 @@ __kernel void makeSphere(__global t_particle *p, int const radius)
 	int		i = get_global_id(0);
 	int x, y, z, w;
 	int		val;
-	double	theta;
-	double phi;
+	float	theta;
+	float phi;
 
 	x = 1 + i;
 	y = 1 + i;
@@ -17,13 +17,13 @@ __kernel void makeSphere(__global t_particle *p, int const radius)
 	y = z;
 	z = w;
     w = w ^ (w >> 19) ^ val ^ (val >> 8);
-	theta = cos((double)w);
-	phi = cos((double)val);
+	theta = cos((float)w);
+	phi = cos((float)val);
 	p[i].pos[0] = radius * theta * phi;
-	theta = cos((double)w);
-	phi = sin((double)val);
+	theta = cos((float)w);
+	phi = sin((float)val);
 	p[i].pos[1] = radius * theta * phi;
-	theta = sin((double)w);
+	theta = sin((float)w);
 	p[i].pos[2] = radius * theta;	
 	p[i].acc[0] = 0.0;
 	p[i].acc[1] = 0.0;
